+++ CHANGELOG.md
@@ -1,4 +1,79 @@
 
+## 4.2.1
+
+
+### Fixed
+
+- Fixed RuntimeException with empty state in onSaveInstanceState [\#160](https://github.com/spacecowboy/NoNonsense-FilePicker/pull/160) ([jangrewe](https://github.com/jangrewe))
+
+## 4.2.0
+
+
+### Added
+
+- Made OK/Cancel strings into library strings for easy overridability [\#157](https://github.com/spacecowboy/NoNonsense-FilePicker/pull/157) ([spacecowboy](https://github.com/spacecowboy))
+
+### Changed
+
+- Update build tools to 26.0.2 and update Dropbox sample to v2 API [\#155](https://github.com/spacecowboy/NoNonsense-FilePicker/pull/155) ([mitchyboy9](https://github.com/mitchyboy9))
+
+## 4.1.0
+
+
+### Added
+
+- added a static helper method for parsing activity results [\#138](https://github.com/spacecowboy/NoNonsense-FilePicker/pull/138) ([spacecowboy](https://github.com/spacecowboy))
+
+
+    Thanks to @F43nd1r for #121
+
+## 4.0.1
+
+
+### Fixed
+
+- Destroy Loader after finish to avoid clearing selections [\#137](https://github.com/spacecowboy/NoNonsense-FilePicker/pull/137) ([spacecowboy](https://github.com/spacecowboy))
+
+## 4.0.0
+
+
+### Breaking changes
+
+- You are now required to define a `FileProvider` in your manifest for the SD-card picker [\#118](https://github.com/spacecowboy/NoNonsense-FilePicker/pull/118) ([spacecowboy](https://github.com/spacecowboy))
+
+
+    Due to recent changes in Android 7.0 Nougat, bare File URIs can no longer be returned in a safe way. This change requires you to add an entry to your manifest to use the included `FilePickerFragment` and change how you handle the results.
+    - You need to add the following to your app's `AndroidManifest.xml`:
+
+    ``` xml
+            <provider
+                android:name="android.support.v4.content.FileProvider"
+                android:authorities="${applicationId}.provider"
+                android:exported="false"
+                android:grantUriPermissions="true">
+                <meta-data
+                    android:name="android.support.FILE_PROVIDER_PATHS"
+                    android:resource="@xml/nnf_provider_paths" />
+            </provider>
+    ```
+    - Then you must change your result handling. Here is a code snippet illustrating the change for a single result (the same applies to multiple results):
+
+    ``` java
+    protected void onActivityResult(int requestCode, int resultCode, Intent intent) {
+        // The URI will now be something like content://PACKAGE-NAME/root/path/to/file
+        Uri uri = intent.getData();
+        // A new utility method is provided to transform the URI to a File object
+        File file = com.nononsenseapps.filepicker.Utils.getFileForUri(uri);
+        // If you want a URI which matches the old return value, you can do
+        Uri fileUri = Uri.fromFile(file);
+        // Do something with the result...
+    }
+    ```
+
+    This change was required in order to fix `FileUriExposedException` being thrown on Android 7.0 Nougat, as reported in [#115](https://github.com/spacecowboy/NoNonsense-FilePicker/issues/115) and [#107](https://github.com/spacecowboy/NoNonsense-FilePicker/issues/107).
+
+    Please see the updated [activity in the sample app](https://github.com/spacecowboy/NoNonsense-FilePicker/blob/master/sample/src/main/java/com/nononsenseapps/filepicker/sample/NoNonsenseFilePicker.java) for more examples.
+
 ## 3.1.0
 
 
@@ -200,7 +275,7 @@
 
     Fixes [\#33](https://github.com/spacecowboy/NoNonsense-FilePicker/issues/33)
 
-## v2.2
+## 2.2.0
 
 
 ### Added
@@ -228,7 +303,7 @@
 
     Fixes [\#29](https://github.com/spacecowboy/NoNonsense-FilePicker/issues/29)
 
-## v2.1
+## 2.1.0
 
 
 ### Changed
@@ -288,7 +363,7 @@
 
     Fixes [\#3](https://github.com/spacecowboy/NoNonsense-FilePicker/issues/3)
 
-## v1.1
+## 1.1.0
 
 
 ### Changed